pyproject.toml
[tool.poetry]
name = "parlay-test"
version = "0.1.0"
description = "Test connections between EC2 and RDS instances."
authors = ["Matt Zhou <mattzh72@mgmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
mysql-connector-python = "^8.0.25"
SQLAlchemy = "^1.4.17"
pandas = "^1.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
